Refactored code due to core and yeoman validator file name changes
