+++ src/commands/falcon/project/create.ts
@@ -18,11 +18,11 @@
  */
 //─────────────────────────────────────────────────────────────────────────────────────────────────┘
 // Imports
-import {core}                   from  '@salesforce/command';             // Salesforce CLI core library
-import {flags}                  from  '@oclif/command';                  // Allows us to define our own custom flags for this command.
-import {GeneratorStatus}        from  '../../../helpers/yeoman-helper'; // Helper object to get status back from Generators after they run.
-import SfdxYeomanCommand        from  '../../../sfdx-yeoman-command';    // Required because this CLI command will launch a Yeoman Generator.
-import {validateLocalPath}      from  '../../../validators/core';       // Core validation function to check that local path values don't have invalid chars.
+import {core}                   from  '@salesforce/command';                // Salesforce CLI core library
+import {flags}                  from  '@oclif/command';                     // Allows us to define our own custom flags for this command.
+import {GeneratorStatus}        from  '../../../helpers/yeoman-helper';     // Helper object to get status back from Generators after they run.
+import SfdxYeomanCommand        from  '../../../sfdx-yeoman-command';       // Required because this CLI command will launch a Yeoman Generator.
+import {validateLocalPath}      from  '../../../validators/core-validator'; // Core validation function to check that local path values don't have invalid chars.
 
 //─────────────────────────────────────────────────────────────────────────────┐
 // SFDX Core library has the ability to import a JSON file with message strings
